% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_metadata.R
\name{format_metadata}
\alias{format_metadata}
\title{Encode metadata table}
\usage{
format_metadata(metadata)
}
\arguments{
\item{metadata}{a metadata tibble to reformat so that it is suitable for transformation into Iramuteq's starred variables.}
}
\value{
metadata formatted as required for Iramuteq
}
\description{
Encode metadata table
}
\examples{
metadata=read.csv(system.file("extdata/fables/metadata.csv",
                              package="textoteR"))
format_metadata(metadata)
}
