% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Gets metadata from specified source.}
\usage{
get_metadata(
  sourcetype,
  from_dir = getwd(),
  filename = "textot.txt",
  encoding = "UTF-8",
  rtibble
)
}
\arguments{
\item{sourcetype}{the type of source ("rtibble","txm" or "iramuteq")}

\item{from_dir}{the corpus directory (contains metadata.csv and multiple .txt files)}

\item{filename}{the name of the Iramuteq-formatted txt file (defaults to "textot.txt")}

\item{encoding}{the encoding of the TXM corpus (defaults to UTF-8)}

\item{rtibble}{the name of the R tibble, if relevant (sourcetype="rtibble")}
}
\value{
a tibble of metadata
}
\description{
Gets metadata from specified source.
}
\examples{
get_metadata(sourcetype="txm",
             from_dir=system.file("extdata/fables",package="textoteR"),
             encoding="UTF-8")
get_metadata(sourcetype="iramuteq",
             from_dir=system.file("extdata",package="textoteR"),
             filename="macron_covid.txt",
             encoding="UTF-8")
get_metadata(sourcetype="rtibble",
             rtibble=LVtweets)
}
