% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iramuteq_to_rtibble.R
\name{iramuteq_to_rtibble}
\alias{iramuteq_to_rtibble}
\title{Converts an Iramuteq corpus file to an R tibble object.}
\usage{
iramuteq_to_rtibble(
  from_dir = NULL,
  filename = "textot.txt",
  encoding = "UTF-8"
)
}
\arguments{
\item{from_dir}{the corpus directory (contains the Iramuteq-formatted file)}

\item{filename}{the name of the Iramuteq-formatted txt file (defaults to "textot.txt")}

\item{encoding}{the encoding of the Iramuteq-formatted file (defaults to UTF-8)}
}
\value{
a R tibble comprising both metadata and texts
}
\description{
Converts an Iramuteq corpus file to an R tibble object.
}
\examples{
rtib=iramuteq_to_rtibble(from_dir=system.file("extdata",package="textoteR"),
                         filename="macron_covid.txt",
                         encoding="UTF-8")
}
