% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iramuteq_to_txm.R
\name{iramuteq_to_txm}
\alias{iramuteq_to_txm}
\title{Converts an Iramuteq corpus file to a TXM corpus directory.}
\usage{
iramuteq_to_txm(
  from_dir = NULL,
  to_dir = "corpus_txm_from_iramuteq",
  filename = "textot.txt",
  encoding = "UTF-8"
)
}
\arguments{
\item{from_dir}{the corpus directory (contains the Iramuteq-formatted file)}

\item{to_dir}{the directory that will contain the output files}

\item{filename}{the name of the Iramuteq-formatted txt file (defaults to "textot.txt")}

\item{encoding}{the encoding of the Iramuteq-formatted file (defaults to UTF-8)}
}
\value{
no R output but new file with name 'filename'
}
\description{
Converts an Iramuteq corpus file to a TXM corpus directory.
}
\examples{
iramuteq_to_txm(from_dir=system.file("extdata",package="textoteR"),
                filename="macron_covid.txt",
                to_dir="macron_covid_corpus")
}
