% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fragments.R
\name{merge_fragments}
\alias{merge_fragments}
\title{Recombines consecutive fragments of natural text into a single text removing redundant parts}
\usage{
merge_fragments(fragments)
}
\arguments{
\item{fragments}{the text fragments to recombine}
}
\value{
the recombined text
}
\description{
Recombines consecutive fragments of natural text into a single text removing redundant parts
}
\examples{
merge_fragments(c("du palais d'un jeune lapin dame belette un beau matin",
                 "belette un beau matin s'empara, c'est une rusée",
                 "c'est une rusée, le maître étant absent,",
                 "ce lui fut chose aisée"))
}
