% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_with.R
\name{replace_with}
\alias{replace_with}
\title{Cleans some patterns in a text by replacing them.}
\usage{
replace_with(text, patterns_to_strings, from = NULL, to = NULL)
}
\arguments{
\item{text}{the character vector to clean.}

\item{from}{the patterns to replace (passed as a vector)}

\item{to}{the patterns to replace (passed as a vector)}

\item{patterns_to_nothing}{the patterns to replace and string replacements (passed as a table with columns "from" and "to")}
}
\value{
the same text with patterns replaced by strings
}
\description{
Cleans some patterns in a text by replacing them.
}
\examples{
replace_with(c("This is an apple and this is an orange.",
               "Here is a banana and this is an apple"),
             from=c("apple","orange"),to=c("pear","kiwi"))
}
