% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_with.R
\name{replace_with_nothing}
\alias{replace_with_nothing}
\title{Cleans some patterns in a text by removing them.}
\usage{
replace_with_nothing(text, patterns_to_nothing)
}
\arguments{
\item{text}{the character vector to clean.}

\item{patterns_to_nothing}{the patterns to replace (either passed as a table with column "from" or passed as a vector)}
}
\value{
the same text with patterns replaced by nothing
}
\description{
Cleans some patterns in a text by removing them.
}
\examples{
replace_with_nothing(c("This is an examplexxx of unclean texxxxt brrr.",
                       "brrrxxxx And anotherxxx one.",
                       "Yet anothxxxer."),
                      c("xxx","brrr"))
}
