% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtibble_to_iramuteq.R
\name{rtibble_to_iramuteq}
\alias{rtibble_to_iramuteq}
\title{Converts an R tibble object to an Iramuteq corpus file.}
\usage{
rtibble_to_iramuteq(
  rtibble,
  to_dir = NULL,
  filename = "textot.txt",
  encoding = "UTF-8"
)
}
\arguments{
\item{rtibble}{the name of the R object containing metadata and text data}

\item{to_dir}{the directory that will contain the output file}

\item{filename}{the name of the Iramuteq-formatted .txt file}

\item{encoding}{encoding of the resulting Iramuteq corpus}
}
\value{
no R output but new file with name 'filename'
}
\description{
Converts an R tibble object to an Iramuteq corpus file.
}
\examples{
rtib= txm_to_rtibble(from_dir=system.file("extdata/fables",
                                          package="textoteR"))
rtibble_to_iramuteq(rtib,
                    filename="textot_fables.txt")
}
