% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtibble_to_txm.R
\name{rtibble_to_txm}
\alias{rtibble_to_txm}
\title{Converts an R tibble object to a TXM corpus directory.}
\usage{
rtibble_to_txm(rtibble, to_dir = NULL, encoding = "UTF-8")
}
\arguments{
\item{rtibble}{the name of the R object containing metadata and text data}

\item{to_dir}{the directory that will contain the output files}

\item{encoding}{the encoding of the resulting TXM corpus (defaults to UTF-8)}
}
\value{
no R output but new directory to_dir
}
\description{
Converts an R tibble object to a TXM corpus directory.
}
\examples{
rtibble_to_txm(LVtweets,
               to_dir="fables_txm")
}
