% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txm_to_iramuteq.R
\name{txm_to_iramuteq}
\alias{txm_to_iramuteq}
\title{Converts a TXM corpus directory to an Iramuteq corpus file.}
\usage{
txm_to_iramuteq(
  from_dir,
  to_dir = NULL,
  filename = "textot.txt",
  encoding = "UTF-8"
)
}
\arguments{
\item{from_dir}{the corpus directory (contains metadata.csv and multiple .txt files)}

\item{to_dir}{the directory that will contain the output file}

\item{filename}{the name of the Iramuteq-formatted txt file (defaults to "textot.txt")}

\item{encoding}{the encoding of the TXM corpus (defaults to UTF-8)}
}
\value{
no R output but new file with name 'filename'
}
\description{
Converts a TXM corpus directory to an Iramuteq corpus file.
}
\examples{
txm_to_iramuteq(from_dir=system.file("extdata/fables",
                                     package="textoteR"),
                filename="data-raw/alltexts_iramuteq.txt")
}
