% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txm_to_rtibble.R
\name{txm_to_rtibble}
\alias{txm_to_rtibble}
\title{Converts a TXM corpus directory to an R tibble object.}
\usage{
txm_to_rtibble(
  from_dir,
  to_dir = NULL,
  filename = "textot.txt",
  encoding = "UTF-8"
)
}
\arguments{
\item{from_dir}{the corpus directory (contains metadata.csv and multiple .txt files)}

\item{to_dir}{the directory that will contain the output file}

\item{filename}{the name of the Iramuteq-formatted txt file (defaults to "textot.txt")}

\item{encoding}{the encoding of the TXM corpus (defaults to UTF-8)}
}
\value{
an R tibble with columns id, text, and all columns from metadata.csv
}
\description{
Converts a TXM corpus directory to an R tibble object.
}
\examples{
txm_to_rtibble(from_dir=system.file("extdata/fables",
                                    package="textoteR"))
}
