% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_words_keep_context.R
\name{unnest_words_keep_context}
\alias{unnest_words_keep_context}
\title{Tokenized by word but produce a segment column to keep the context of the words.}
\usage{
unnest_words_keep_context(data, input, nwords = 20)
}
\arguments{
\item{data}{data to tokenize (as provided to tidytext::unnest_tokens)}

\item{input}{the column of the data that contains the text to be tokenized}

\item{nwords}{the number of words that constitute a segment}
}
\value{
the tokenized data with a segment column as word context.
}
\description{
Tokenized by word but produce a segment column to keep the context of the words.
}
\examples{
corpus_data=iramuteq_to_rtibble(from_dir="data-raw",filename="alltexts_iramuteq.txt")
unnest_words_keep_context(corpus_data,input="text",nwords=15)
}
