% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_iramuteq.R
\name{write_to_iramuteq}
\alias{write_to_iramuteq}
\title{Writes metadata and textdata as an Iramuteq corpus file .}
\usage{
write_to_iramuteq(
  metadata,
  textdata,
  to_dir = NULL,
  filename = "textot.txt",
  encoding = "UTF-8"
)
}
\arguments{
\item{metadata}{metadata R tibble}

\item{textdata}{textdata R tibble}

\item{to_dir}{the directory that will contain the output files}

\item{filename}{the name of the Iramuteq-formatted txt file (defaults to "textot.txt")}

\item{encoding}{the encoding of the Iramuteq-formatted file (defaults to UTF-8)}
}
\value{
no R output but new file with name filename
}
\description{
Writes metadata and textdata as an Iramuteq corpus file .
}
\examples{
metadata=get_metadata(sourcetype="Iramuteq",
                      from_dir="data-raw",
                      filename="macron_covid.txt")
textdata=get_textdata(sourcetype="Iramuteq",
                      from_dir="data-raw",
                      filename="macron_covid.txt")
write_to_iramuteq(metadata,
                  textdata,
                  to_dir="outputs",
                  encoding="UTF-8")
}
