% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_txm.R
\name{write_to_txm}
\alias{write_to_txm}
\title{Writes metadata and textdata as a TXM corpus directory (with UTF-8 encoding).}
\usage{
write_to_txm(metadata, textdata, to_dir = "corpus_txm")
}
\arguments{
\item{metadata}{metadata R tibble}

\item{textdata}{textdata R tibble}

\item{to_dir}{the directory that will contain the output files}
}
\value{
no R output but new directory to_dir
}
\description{
Writes metadata and textdata as a TXM corpus directory (with UTF-8 encoding).
}
\examples{
metadata=get_metadata(sourcetype="Iramuteq",
                      from_dir="data-raw",
                      filename="textot.txt")
textdata=get_textdata(sourcetype="Iramuteq",
                      from_dir="data-raw",
                      filename="textot.txt")
write_to_txm(metadata,
             textdata,
             to_dir="corpus_txm")
}
