% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textoteR.R
\docType{package}
\name{textoteR}
\alias{textoteR-package}
\alias{textoteR}
\title{\code{textoteR} package}
\description{
A package that converts text corpora sources from/to possible formats: TXM, Iramuteq, and R tibble.
See the README on \href{https://github.com/lvaudor/mixr#readme}{GitHub}
}
